requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
pymongo[srv]==4.12.0
python-docx==1.1.2
authlib==1.4.1
httpx==0.28.0
bcrypt==4.3.0
PyJWT==2.10.0
pydantic[email]==2.10.0
pydantic-settings==2.7.0
slowapi==0.1.9
python-multipart==0.0.20
itsdangerous==2.2.0
pytest==8.3.0
pytest-asyncio==0.25.0
